Adding a macro for sdjagent.jar running inside agent container
Currently, we make usage of the macro 'java_running_sdjagent', which
relies on the process command line argument.  We recently noticed that
in certain cases, in a containerized environment, the process command
line does not contain sdjagent.jar. Introducing the
'java_in_agent_container' macro and make use on the two rules that
relied solely on the command line macro to avoid alerts.

Adding a comment within the rule, to explain the need and usage of
both macros.